+++ CHANGELOG.md
@@ -8,8 +8,9 @@ _general description_
 
 - **added** - random number generator
 - **changed** - better concept names and organization
-- **added** - vector algebra
 - **added** - constexpr for implementations
+- **added** - vector algebra
+- **added** - matrix algebra
 - **todo** - global/local job queues
 
 #### [0.4.1] job system (_21 jun 22_)